requirements.txt
APScheduler==3.6.1
astroid==2.2.5
autopep8==1.4.4
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
dataclasses==0.6
Django==2.2.4
django-guardian==2.1.0
django-referrer-policy==1.0
django-rest-swagger==2.2.0
django-stubs==0.10.2
djangorestframework==3.9.3
djangorestframework-simplejwt==4.3.0
djangorestframework-stubs==0.4.2
ecdsa==0.13.2
future==0.17.1
idna==2.8
isort==4.3.21
itypes==1.1.0
Jinja2==2.10.1
lazy-object-proxy==1.4.2
MarkupSafe==1.1.1
mccabe==0.6.1
mypy==0.701
mypy-extensions==0.4.1
openapi-codec==1.3.2
peewee==3.11.2
psycopg2-binary==2.8.2
pyasn1==0.4.7
pycodestyle==2.5.0
pyfiglet==0.8.post1
PyJWT==1.7.1
pylint==2.3.1
python-decouple==3.1
python-jose==3.0.1
pytz==2019.1
requests==2.22.0
rsa==4.0
simplejson==3.16.0
six==1.12.0
sqlparse==0.3.0
typed-ast==1.3.5
typing-extensions==3.7.2
tzlocal==2.0.0
uritemplate==3.0.0
urllib3==1.25.3
wrapt==1.11.2
zathura==0.0.5.8b1
